move cores3_audio_codec.cc to boards
