+++ main/boards/m5stack-core-s3/m5stack_core_s3.cc
@@ -1,5 +1,5 @@
 #include "wifi_board.h"
-#include "audio_codecs/cores3_audio_codec.h"
+#include "cores3_audio_codec.h"
 #include "display/lcd_display.h"
 #include "application.h"
 #include "config.h"
